Refactor the configuration handling
Move the configuration state from the UI to the ApplicationState to avoid having
to maintain multiple copies of the configuration.

Also make the data flow clearer around the ConfigurationWindow by reloading
the ApplicationState configuration before showing the dialog instead of relying
on the Configuration global acting both as input and output leaving the purpose
of some bindings being active or not quite unclear.

Remove the toml_edit serialization. The purpose was to allow keeping comments
and spacing from the file on disk, but the API is making that one goal very
cumbersome to achieve since I'd need to manually carry the Decor of each
value around when writing in the document.
Just use serde with toml instead and assume that most edits will be done by
the UI until I can know if this is really helpful or not.